More low-hanging fixes (WIP)
