+++ include/ccf/claims_digest.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "crypto/hash_provider.h"
+#include "ccf/crypto.h"
 
 namespace ccf
 {
