+++ include/ccf/indexing/strategies/visit_each_entry_in_map.h
@@ -4,6 +4,7 @@
 
 #include "ccf/indexing/strategy.h"
 #include "indexing/indexing_types.h"
+#include "kv/store.h"
 
 namespace ccf::indexing::strategies
 {
